Construct dask arrays with explicit meta
It can be tricky to infer the dtype of a dask array since it is lazy.
In addition, the array type is not necessarily np.ndarray, so getting
a handle on that is helpful too. Dask conveys this info with a `meta`
parameter: an empty array of the expected dtype and class (and ndim?).

If you don't provide `meta` during array creation, I had expected dask
to use something like `np.array((), dtype)` since we do provide `dtype`.
Alas, no. It computes an empty slice of the dask array just to figure
out the dtype and class all over again.

In the case of S3ChunkStore, this fetches actual chunks from the store.
This is repeated for vis, flags, weights and weights_channel. It adds
several seconds to the duration of `katdal.open` for no good reason.

Save this time by explicitly providing `meta` to `get_dask_array`.
It looks like `put_dask_array` is not affected to the same extent,
but add `meta` there too just to be sure.