SPR1-2655 Check rclone and its version
Only continue if rclone >= 1.56 is installed. The rclone versions
between 1.52 and 1.56 fleshed out the "memory-only config" scenario
which we rely on. Before that, RCLONE_CONFIG="" looked for a config file
in the current directory and had to be replaced by the awkward and
platform-dependent /dev/null to disable the config file. More seriously,
the --files-from filter could not read from stdin, which complicates
the selection mechanism.

Add `packaging` to the requirements, since we use it to parse and
compare the rclone version. It is already needed by pytest and therefore
quite a benign addition.